requirements.txt
google-genai>=1.0.0
twilio>=9.0.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
piazza-api>=0.9.0
flask>=3.0.0
tzdata
